requirements.txt
click==6.7
enum-compat==0.0.2
enum34==1.1.6
eventlet==0.21.0
Flask==1.0
Flask-SocketIO==2.8.6
greenlet==0.4.15
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
pyinotify==0.9.6
python-dateutil==2.6.0
python-engineio==3.8.2.post1
python-socketio==1.7.4
redis==2.10.5
RPi.GPIO==0.6.3
six==1.10.0
spidev==3.2
Werkzeug==0.15.3
